requirements.txt
aiofiles==24.1.0
aiosmtplib==2.0.2
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
fastapi-cli==0.0.5
fastapi-mail==1.4.1
Flask==3.0.3
fonttools==4.54.1
fprint==1.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
imageio==2.35.1
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
kiwisolver==1.4.7
lazy_loader==0.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
networkx==3.4.1
numpy==2.1.2
opencv-python==4.10.0.84
orjson==3.10.7
packaging==24.1
passlib==1.7.4
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
PyYAML==6.0.2
rich==13.9.2
rsa==4.9
scikit-image==0.24.0
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tifffile==2024.9.20
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.31.1
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
Werkzeug==3.0.4
